:sparkles: util/yaml: generic Cluster API yaml parser for clusterctl
